+++ src/services/PrivateKeyService.js
@@ -1,25 +1,44 @@
+const bip39 = require('bip39')
 const ethereumUtil = require('ethereumjs-util')
 const fs = require('fs')
+
+const hdkey = require('ethereumjs-wallet/hdkey')
 const { promisify } = require('util')
 
 const readFileAsync = promisify(fs.readFile)
 
 const { BaseError } = require('../utils/errors')
+const log = require('../logger')
+const PrivateKeyValidator = require('../validators/PrivateKeyValidator')
+
+const privateKeyValidatorInstance = new PrivateKeyValidator(log)
 
 class EthereumAddressError extends BaseError {}
 class InvalidPrivateKeyError extends BaseError {}
 class InvalidPrivateKeyFile extends BaseError {}
+class InvalidMnemonicError extends BaseError {}
 
+const HD_PATH = "m/44'/60'/0/0"
 const encoding = 'utf8'
 /**
  * Class representing a service that manage private key.
  */
 class PrivateKeyService {
-  constructor(logger) {
+  constructor(privateKeyValidator = privateKeyValidatorInstance, logger = log, hdPath = HD_PATH) {
+    if (!privateKeyValidator) {
+      throw new TypeError(`Invalid "privateKeyValidator" value: ${privateKeyValidator}`)
+    }
+    this.privateKeyValidator = privateKeyValidator
+
     if (!logger) {
       throw new TypeError(`Invalid "logger" value: ${logger}`)
     }
     this.log = logger.child({ module: this.constructor.name })
+
+    if (!hdPath) {
+      throw new TypeError(`Invalid "hdPath" value: ${hdPath}`)
+    }
+    this.hdPath = hdPath
   }
 
   /**
@@ -66,6 +85,28 @@ class PrivateKeyService {
       throw new EthereumAddressError(err)
     }
   }
+
+  /**
+   * It gets the private key from mnemonic
+   *
+   * @param {String} mnemonic The mnemonic.
+   *
+   * @throws {InvalidMnemonicError} If the menomic is not valid.
+   */
+  getPrivateKeyFromMnemonic(mnemonic) {
+    try {
+      const validMnemonic = this.privateKeyValidator.validateMnemonic(mnemonic)
+      const hdwallet = hdkey.fromMasterSeed(bip39.mnemonicToSeed(validMnemonic))
+      const wallet = hdwallet.derivePath(this.hdPath).getWallet()
+      const privateKey = wallet._privKey.toString('hex')
+      this.log.debug({ fn: 'getPrivateKeyFromMnemonic' }, 'Retrieve private key from mnemonic.')
+      return privateKey
+    } catch (err) {
+      this.log.error({ fn: 'getPrivateKeyFromMnemonic', err },
+        'Get address from private key.')
+      throw new InvalidMnemonicError(err)
+    }
+  }
 }
 
-module.exports = { PrivateKeyService, InvalidPrivateKeyError, InvalidPrivateKeyFile }
+module.exports = { PrivateKeyService, InvalidPrivateKeyError, InvalidPrivateKeyFile, InvalidMnemonicError }